use the same method for update and create

+++ internal/domain/models/transaction.go
@@ -7,11 +7,11 @@ import (
 )
 
 type TransactionBalance struct {
-	Value    int `bson:"value" json:"value"`
-	Parts    int `bson:"parts" json:"parts"`       // increase
-	Labor    int `bson:"labor" json:"labor"`       // increase
-	Discount int `bson:"discount" json:"discount"` // decrease
-	Interest int `bson:"interest" json:"interest"` // increase
+	Value    float64 `bson:"value" json:"value"`
+	Parts    float64 `bson:"parts" json:"parts"`       // increase
+	Labor    float64 `bson:"labor" json:"labor"`       // increase
+	Discount float64 `bson:"discount" json:"discount"` // decrease
+	Interest float64 `bson:"interest" json:"interest"` // increase
 }
 
 type TransactionRepeatSettings struct {
